+++ src/classes/flow-producer.ts
@@ -163,10 +163,20 @@ export class FlowProducer extends EventEmitter {
     const client = await this.connection.client;
     const multi = client.multi();
 
+    const parentOpts = flow?.opts?.parent;
+    const parentKey = getParentKey(parentOpts);
+    const parentDependenciesKey = parentKey
+      ? `${parentKey}:dependencies`
+      : undefined;
+
     const jobsTree = this.addNode({
       multi,
       node: flow,
       queuesOpts: opts?.queuesOptions,
+      parent: {
+        parentOpts,
+        parentDependenciesKey,
+      },
     });
 
     await multi.exec();