+++ tests/test_clean.ts
@@ -450,7 +450,7 @@ describe('Cleaner', () => {
                       {
                         parent: {
                           id: job.id,
-                          queue: `${job.prefix}:${job.queueName}`,
+                          queue: job.queueQualifiedName,
                         },
                       },
                     );
@@ -477,6 +477,8 @@ describe('Cleaner', () => {
                       });
                       step = Step.Finish;
                       return Step.Finish;
+                    } else {
+                      job.autoComplete = true;
                     }
                     break;
                   }