+++ tests/test_connection.ts
@@ -97,6 +97,20 @@ describe('connection', () => {
         );
         await connection.disconnect();
       });
+
+      describe('when using nodes provides an array of strings as hosts', async () => {
+        it('throws an error', async () => {
+          const connection = new IORedis.Cluster(
+            ['localhost', 'https://upstash.io'],
+            {},
+          );
+
+          expect(() => new QueueBase(queueName, { connection })).to.throw(
+            'BullMQ: Upstash is not compatible with BullMQ.',
+          );
+          await connection.disconnect();
+        });
+      });
     });
   });